split JobRunner out of model manager
